+++ test/test_runtime_contract.rb
@@ -27,6 +27,7 @@ plasmoid = project_root.join("kde/plasmoids/org.xnix.compatibilitycenter/metadat
   GetEngineCatalog
   GetRunPlan
   GetApplicationStateRoot
+  GetCompatibilityPackageSource
   GetBackendBinding
   GetRepairPlan
   GetTestPlan